Http Condition: close http connection once finished
The Http condition will open a http connection and check that the
status code corresponds with a success code.

Unfortunately, Http condition does not then close the connection
afterwards and this has been found to stall builds when connecting
through web proxies.

This patch closes the http connection once we've finished with it.

Signed-off-by: Steve Capper <steve.capper@linaro.org>